Fix CHECK_ENOUGH in fb_unserialize
Summary:
CHECK_ENOUGH was casting things which might be int64_t to
int.

Reviewed By: mofarrell

Differential Revision: D16651507

+++ hphp/runtime/ext/fb/ext_fb.cpp
@@ -569,7 +569,7 @@ Variant HHVM_FUNCTION(fb_compact_serialize, const Variant& thing) {
 
 /* Check if there are enough bytes left in the buffer */
 #define CHECK_ENOUGH(bytes, pos, num) do {                      \
-    if ((int)(bytes) > (int)((num) - (pos))) {                  \
+    if ((int64_t)(bytes) > (int64_t)((num) - (pos))) {          \
       return FB_UNSERIALIZE_UNEXPECTED_END;                     \
     }                                                           \
   } while (0)
